requirements/arxiv.txt
arxiv
pyyaml
requests
pyTelegramBotAPI